Raise scala version to 3.7.3
