+++ build.sbt
@@ -1,4 +1,4 @@
-val scala3Version = "3.3.0"
+val scala3Version = "3.7.3"
 
 lazy val root = project
   .in(file("."))
@@ -7,6 +7,10 @@ lazy val root = project
     version := "0.1.0-SNAPSHOT",
 
     scalaVersion := scala3Version,
+    scalacOptions ++= Seq(
+      "-deprecation",
+      "-feature",
+    ),
 
     libraryDependencies ++= Seq(
       "org.scalactic" %% "scalactic" % "3.2.16",
