+++ src/main/scala/tyes/compiler/TermIRGenerator.scala
@@ -6,10 +6,6 @@ import tyes.compiler.target.TargetCodeTypeRef
 import tyes.model.Term
 import tyes.model.Type
 
-private val TCN = TargetCodeNode
-private val TCP = TargetCodePattern
-private val TCTypeRef = TargetCodeTypeRef
-
 class TermIRGenerator(
   private val typeIRGenerator: TypeIRGenerator
 ):
@@ -17,6 +13,7 @@ class TermIRGenerator(
   def generate(term: Term, codeEnv: TargetCodeEnv = TargetCodeEnv()): TargetCodeNode = term match {
     case Term.Constant(value: Int) => TCN.Integer(value)
     case Term.Constant(value: String) => TCN.Text(value)
+    case Term.Constant(_) => ???
     case v: Term.Variable => codeEnv(v)
     case Term.Function(name, args*) => 
       TCN.ADTConstructorCall(
@@ -29,6 +26,7 @@ class TermIRGenerator(
   def generatePattern(term: Term): TargetCodePattern = term match {
     case Term.Constant(value: Int) => TCP.Integer(value)
     case Term.Constant(value: String) => TCP.Text(value)
+    case Term.Constant(_) => ???
     case Term.Variable(name) => TCP.Var(name)
     case Term.Function(name, args*) => 
       TCP.ADTConstructor(
