+++ src/main/scala/tyes/compiler/aliases.scala
@@ -1,0 +1,18 @@
+package tyes.compiler
+
+import tyes.compiler.target.TargetCodeNode
+import tyes.compiler.target.TargetCodeTypeRef
+import tyes.compiler.target.TargetCodePattern
+import tyes.compiler.target.TargetCodeADTConstructor
+import tyes.compiler.target.TargetCodeDecl
+import tyes.compiler.target.TargetCodeForCursor
+
+private type TCN = TargetCodeNode
+private type TCTypeRef = TargetCodeTypeRef
+
+private val TCN = TargetCodeNode
+private val TCP = TargetCodePattern
+private val TCTypeRef = TargetCodeTypeRef
+private val TCD = TargetCodeDecl
+private val TCADTConstructor = TargetCodeADTConstructor
+private val TCFC = TargetCodeForCursor
