+++ src/main/scala/tyes/model/Term.scala
@@ -18,7 +18,7 @@ enum Term extends terms.TermOps[Term, Any](TermBuilder):
 
   override def matches(otherTerm: Term): Option[Map[String, Term]] = 
     ifBothTypes(otherTerm) { (thisType, otherType) =>
-      thisType.matches(otherType).map(_.mapValues(Term.Type.apply).toMap)
+      thisType.matches(otherType).map(_.view.mapValues(Term.Type.apply).toMap)
     }.getOrElse { 
       super.matches(otherTerm) 
     }
@@ -42,7 +42,7 @@ enum Term extends terms.TermOps[Term, Any](TermBuilder):
 
   override def unifies(otherTerm: Term): Option[Map[String, Term]] =
     ifBothTypes(otherTerm) { (thisType, otherType) =>
-      thisType.unifies(otherType).map(_.mapValues(Term.Type.apply).toMap)
+      thisType.unifies(otherType).map(_.view.mapValues(Term.Type.apply).toMap)
     }.getOrElse { 
       super.unifies(otherTerm) 
     }
